feat(cuda): add CUDA kernel support and Narval A100 benchmark
- Fix CUDA kernel compilation: replace INFINITY with 1.0/0.0 (CuPy
  RawKernel device code has no access to math.h)
- Fix dtype mismatches for CUDA path (float64 kernel output vs float32
  PyTorch tensors): fold_losses, weights_gpu, scores tensor
- Add cuda_kernel backend to variance suite configs
- Add Narval SLURM scripts (setup, data download, quick/full bench)
- Add traceback logging to runner for easier debugging
- A100 results: 20-82x speedup vs CPU, ~30% less VRAM than torch

+++ autoreject/kernels/batched_scoring.py
@@ -122,7 +122,7 @@ def batched_consensus_score(X_train_interp_gpu: Any, median_gpu: Any,
 
     # Transfer back and apply validity mask
     scores = torch_module.full((n_consensus,), float('-inf'),
-                               device=rmse.device)
+                               device=rmse.device, dtype=rmse.dtype)
     valid_t = torch_module.tensor(valid_mask, device=rmse.device)
     scores[valid_t] = -rmse[valid_t]
